NHSO-0000: Added dry_run to release cleanup script

+++ nhsapp-chart/utils/release_management/create_script.sh
@@ -7,6 +7,16 @@ function die () {
   exit 1
 } 
 
+dry_run=false
+
+while getopts 'd' opt; do
+    case "$opt" in
+        d) dry_run=true ;;
+        *) echo 'error in command line parsing' >&2
+           exit 1
+    esac
+done
+
 [ -z "$NAMESPACE" ] && die "Namespace not specified, exiting..."
 [ -z "$RELEASE_VERSION" ] && die "Release version not specified, exiting..."
 
@@ -16,10 +26,18 @@ RESOURCES=$(kubectl get all -n "${NAMESPACE}" --selector="version!=${RELEASE_VER
 
 # Horizontal Pod Autoscalers and Pod Disruption Budgets
 while IFS= read -r r; do
+  [ -z "$r" ] && die "Resources not defined, exiting..."
   echo "kubectl delete --namespace ${NAMESPACE} ${r} --wait=false" >> prune_releases.sh
 done <<<"${RESOURCES}"
 
 chmod +x prune_releases.sh
 
-echo "Please review the following script before approving:"
-cat prune_releases.sh
+if [[ $dry_run == true ]]; then
+  echo "Please review the following script before approving:"
+  cat prune_releases.sh
+elif [[ $dry_run == false ]]; then
+  ./prune_releases.sh
+else
+  die "Error setting dry_run variable, exiting..."
+fi
+
